Refactoring Runtime.Api - SuggestionAction
Moving `SuggestionAction` out of the `Runtime` class.
